requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
motor==3.4.0        
bcrypt==4.1.2       
python-jose==3.3.0  
pymongo==4.6.3
pytest
pytest-asyncio
python-dotenv
